Move more helper functions to utilities.cpp
